# docs/swagger.yaml
definitions:
  model.CreateCustomerRequest:
    properties:
      document:
        type: string
      income_range:
        type: string
      name:
        type: string
      risk_level:
        type: string
      score:
        type: integer
    required:
    - document
    - name
    - risk_level
    - score
    type: object
  model.Customer:
    properties:
      created_at:
        type: string
      document:
        type: string
      id:
        type: string
      income_range:
        type: string
      name:
        type: string
      risk_level:
        type: string
      score:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  model.UpdateStatusRequest:
    properties:
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /customers:
    post:
      consumes:
      - application/json
      description: Validates and saves a new customer registration with an initial
        ACTIVE status.
      parameters:
      - description: Customer registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create a new customer record
      tags:
      - customers
  /customers/{id}/status:
    patch:
      consumes:
      - application/json
      description: Partially modifies a customer profile status state (ACTIVE, INACTIVE,
        UNDER_REVIEW).
      parameters:
      - description: Customer UUID
        in: path
        name: id
        required: true
        type: string
      - description: Target status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Update customer status
      tags:
      - customers
swagger: "2.0"